requirements.txt
Flask==3.1.0
requests==2.32.3
numpy==1.26.4
Pillow==11.1.0
onnxruntime==1.21.0
gunicorn==22.0.0
